+++ swarm.cabal
@@ -81,7 +81,8 @@ common ghc2021-extensions
 
 library
     import:           stan-config, common, ghc2021-extensions
-    exposed-modules:  Data.BoolExpr.Simplify
+    exposed-modules:  Brick.Widget.List.Skippable
+                      Data.BoolExpr.Simplify
                       Swarm.App
                       Swarm.DocGen
                       Swarm.Game.Failure
